requirements.txt
argh==0.26.2
h5py==3.5.0
matplotlib==3.4.3
numpy==1.21.3
pathfinding==1.0.1
pyzmq==22.3.0
scipy==1.7.1
seaborn==0.11.2
tqdm==4.62.3
